projects/image-link/backend/pyproject.toml
[project]
name = "imagelink-backend"
version = "0.1.0"
description = "ImageLink API"
requires-python = ">=3.10"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi>=0.115.0",
    "python-dotenv>=1.2.2",
    "supabase>=2.28.3",
    "uvicorn[standard]>=0.32.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
